feat(docker): statically allocate a network and IP addresses
This basically makes Petri in control of the IP address management
instead of docker. This makes it easier to get static IP addresses
even if the container is stopped. Before this, the Docker provider
would provide DNS names instead of IP addresses, which is
undesirable due to CometBFT/the SDK not having good support for DNS
hostnames.