fix(provider): improve recognition of exit codes
currently, both DO and Docker providers do not wait for the container to
exit until returning an exit code. this introduces a condition, where
the container hasn't exited yet (e.g. with an error), but we already
fetch it's stdout and exit code (which at that point is 0). this commit
makes the providers wait for the exec container to finish and only then
returns the response.

(cherry picked from commit 4a670527d3ccfa408a44530125dd386fda0151af)

# Conflicts:
#	core/provider/digitalocean/task.go
#	core/provider/docker/task.go
#	cosmos/node/init.go